dgh-platform/backend/api-gateway/requirements.txt
anyio==4.9.0
asgiref==3.9.1
async-timeout==5.0.1
certifi==2025.7.14
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
packaging==25.0
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
